"Refactor BookEntity 'Genre' to support multiple genres"
Changes have been made to the BookEntity and the associated GraphQL type to refactor the 'Genre' field from a string to an IEnumerable<GenreEntity>. This was necessary in order to allow books to include multiple genres. This involved altering EntityToTypeMapping for the BookEntity, as well as the relevant Entity classes and configurations. Redundant spaces in 'Hot