Refactor attribute usage and remove unnecessary usages
Remove unnecessary usages in import statements for several files in the Persistence and Worker directories. Bogus, System.Diagnostics, System.ComponentModel, System.Linq.Expressions, and Microsoft.EntityFrameworkCore.Infrastructure are no longer in use in the given files, hence they are removed. Additionally, in the User class and the Book class in the GraphQL/Types directory, the attribute [IsProjected(true)] has been simplified to [IsProjected], removing the need to specify the value when it's the default. This commit also corrects the Author mapping in the BookEntityExtension class, ensuring the correct function is called on the Author object. This refactor improves readability and efficiency of the code.